split runtime trace tests
