+++ internal/runtime/trace_load_test.go
@@ -1,0 +1,83 @@
+package runtime
+
+import (
+	"errors"
+	"os"
+	"path/filepath"
+	"strings"
+	"testing"
+)
+
+func TestLoadTrace(t *testing.T) {
+	trace, err := loadTraceFromContent(t, `{"kind":"resolve","module":"`+lodashMapModule+`","resolved":"file:///repo/node_modules/lodash/map.js"}`+"\n"+`{"kind":"require","module":"@scope/pkg/lib","resolved":"/repo/node_modules/@scope/pkg/lib/index.js"}`+"\n")
+
+	if err != nil {
+		t.Fatalf(loadTraceErrFmt, err)
+	}
+	if trace.DependencyLoads["lodash"] != 1 {
+		t.Fatalf("expected lodash load count=1, got %d", trace.DependencyLoads["lodash"])
+	}
+	if trace.DependencyLoads[scopePkgDependency] != 1 {
+		t.Fatalf("expected %s load count=1, got %d", scopePkgDependency, trace.DependencyLoads[scopePkgDependency])
+	}
+	if got := trace.DependencyModules["lodash"][lodashMapModule]; got != 1 {
+		t.Fatalf("expected lodash module count 1, got %d", got)
+	}
+	if got := trace.DependencySymbols["lodash"][lodashMapModule+"\x00map"]; got != 1 {
+		t.Fatalf("expected lodash symbol count 1, got %d", got)
+	}
+}
+
+func TestLoadTraceInvalidLine(t *testing.T) {
+	if _, err := loadTraceFromContent(t, "{not-json}\n"); err == nil {
+		t.Fatalf("expected parse error for invalid NDJSON")
+	}
+}
+
+func TestLoadTraceScannerErrTooLong(t *testing.T) {
+	tooLong := strings.Repeat("x", 80*1024)
+	_, err := loadTraceFromContent(t, tooLong)
+	if err == nil {
+		t.Fatalf("expected scanner error for oversized line")
+	}
+}
+
+func TestLoadTraceParseErrorIncludesLineNumber(t *testing.T) {
+	_, err := loadTraceFromContent(t, "{\"module\":\"ok\"}\n{not-json}\n")
+	if err == nil {
+		t.Fatalf("expected parse error")
+	}
+	if !strings.Contains(err.Error(), "line 2") {
+		t.Fatalf("expected line number in parse error, got %v", err)
+	}
+}
+
+func TestLoadTraceSkipsBlankLines(t *testing.T) {
+	trace, err := loadTraceFromContent(t, "\n   \n{\"module\":\""+lodashMapModule+"\"}\n")
+	if err != nil {
+		t.Fatalf(loadTraceErrFmt, err)
+	}
+	if got := trace.DependencyLoads["lodash"]; got != 1 {
+		t.Fatalf("expected lodash load count 1, got %d", got)
+	}
+}
+
+func TestLoadTraceMissingFileError(t *testing.T) {
+	_, err := Load(filepath.Join(t.TempDir(), "missing.ndjson"))
+	if err == nil {
+		t.Fatalf("expected missing-file error")
+	}
+	if !errors.Is(err, os.ErrNotExist) {
+		t.Fatalf("expected os.ErrNotExist, got %v", err)
+	}
+}
+
+func TestLoadTraceSkipsEventsWithoutDependencies(t *testing.T) {
+	trace, err := loadTraceFromContent(t, "{\"module\":\"./local\"}\n{\"resolved\":\"/repo/src/index.js\"}\n")
+	if err != nil {
+		t.Fatalf(loadTraceErrFmt, err)
+	}
+	if len(trace.DependencyLoads) != 0 || len(trace.DependencyModules) != 0 || len(trace.DependencySymbols) != 0 {
+		t.Fatalf("expected dependency-free events to be ignored, got %#v", trace)
+	}
+}
